backend/pyproject.toml
[project]
name = "dinedi-backend"
version = "1.0.0"
description = "API de suppression de watermarks NotebookLM — PDF, PPTX, PNG, JPG"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.0",
    "img2pdf>=0.6.3",
    "numpy>=2.4.4",
    "opencv-python-headless>=4.13.0.92",
    "pillow>=12.2.0",
    "pydantic>=2.13.2",
    "pymupdf>=1.27.2.2",
    "python-multipart>=0.0.26",
    "python-pptx>=1.0.2",
    "uvicorn>=0.44.0",
]

 